backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
cohere==4.37
python-dotenv==1.0.0
aiohttp==3.9.1
tweepy==4.14.0
tavily-python==0.2.3
pydantic==2.5.3 